api/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.18.0
SQLAlchemy>=1.4
asyncpg>=0.27.0
python-dotenv>=1.0.0
aio-pika>=8.3.0
